Replication, MySQL server_id system variable is being used as an int instead of a long which is causing an error sometimes. Backported to: Reviewed by: Allen
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@13044 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormReplicationSetup.cs
@@ -145,7 +145,7 @@ namespace OpenDental {
 		}
 
 		private void butTest_Click(object sender,EventArgs e) {
-			int server_id=ReplicationServers.Server_id;
+			long server_id=ReplicationServers.Server_id;
 			string msg="";
 			if(server_id==0) {
 				msg="server_id not set for this server.\r\n\r\n";
